评教系统最终版/评教系统管理端/backend_8fMBP/backend/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
sqlalchemy==2.0.45
python-multipart==0.0.20
websockets==16.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.0
email-validator==2.2.0
httpx==0.28.1
requests==2.32.3

# 文件解析依赖
python-docx==1.1.2
PyPDF2==3.0.1
python-pptx==1.0.2

# 加密和哈希
cryptography==44.0.0

# 测试依赖
pytest==8.3.4
hypothesis==6.92.1
